requirements.txt
fastapi==0.58.0
pynamodb==4.3.2
pyyaml==5.3.1
requests==2.24.0
uvicorn==0.11.7